+++ CHANGELOG.md
@@ -32,3 +32,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `superForm.isTainted(path?)`
 - `superValidate.options.allowFiles` which makes `setError` and `message` remove files automatically by default. Use `removeFiles` and `failAndRemoveFiles` in other cases.
 - SuperDebug now displays `File` and `FileList`.
+- All proxies can now take the whole `superForm` object with an extra `taint` option.